requirements.txt
adagio==0.2.6
alembic==1.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
ansi2html==1.9.2
anyio==4.12.0
aplr==10.19.3
appdirs==1.4.4
appnope @ file:///home/conda/feedstock_root/build_artifacts/appnope_1733332318622/work
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.2
catboost==1.1.1
category-encoders==2.6.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
choreographer==1.2.1
click==8.1.8
cloudpickle==3.1.2
colour==0.1.5
comm @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_comm_1753453984/work
contourpy==1.3.0
cryptography==46.0.3
cycler==0.12.1
Cython==3.2.2
dacite==1.9.2
dash==3.3.0
dash-bootstrap-components==2.0.4
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_auth==2.3.0
dash_cytoscape==1.0.2
databricks-sdk==0.74.0
debugpy @ file:///opt/miniconda3/conda-bld/debugpy_1757067185699/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
deprecation==2.1.0
dill==0.4.0
distro==1.9.0
docker==7.1.0
dtreeviz==2.2.2
dynaconf==3.2.12
evidently==0.7.17
exceptiongroup==1.3.1
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1745502089858/work
explainerdashboard==0.4.8
fairlearn==0.7.0
Faker==37.12.0
fastapi==0.124.2
fastjsonschema==2.21.2
filelock==3.19.1
filetype==1.2.0
Flask==3.1.2
Flask-WTF==1.2.2
flask_simplelogin==0.2.0
fonttools==4.60.2
fsspec==2025.10.0
fugue==0.9.3
gevent==25.9.1
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.43.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
graphviz==0.21
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
ImageHash==4.3.1
imbalanced-learn==0.12.4
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
interpret==0.6.9
interpret-core==0.6.9
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1754352890318/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1701831663892/work
ipywidgets==8.1.8
iterative-telemetry==0.0.10
itsdangerous==2.2.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-dash==0.4.2
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1748333051527/work
jupyterlab_widgets==3.0.16
kaleido==1.2.0
kiwisolver==1.4.7
kmodes==0.12.2
lightgbm==4.6.0
litestar==2.18.0
litestar-htmx==0.5.0
llvmlite==0.43.0
logistro==2.0.1
Mako==1.3.10
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.7.5
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mdurl==0.1.2
minify_html==0.18.1
mlflow==3.1.4
mlflow-skinny==3.1.4
mlxtend==0.23.4
mpmath==1.3.0
msgspec==0.20.0
multidict==6.7.0
multimethod==1.12
multipart==1.3.0
multiprocess==0.70.18
mypy_extensions==1.1.0
nbformat==5.10.4
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.2.1
nltk==3.9.2
numba==0.60.0
numpy==2.0.2
openai==2.9.0
opentelemetry-api==1.39.0
opentelemetry-proto==1.39.0
opentelemetry-sdk==1.39.0
opentelemetry-semantic-conventions==0.60b0
orjson==3.11.5
oyaml==1.0
packaging==25.0
pandas==2.3.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
patsy==1.0.2
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
phik==0.12.5
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==11.3.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1746710438/work
plotly==5.24.1
plotly-resampler==0.11.0
pluggy==1.6.0
pmdarima==2.0.4
polyfactory==3.1.0
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1744724089886/work
protobuf==6.33.2
psutil==7.1.3
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
puremagic==1.30
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycaret==3.3.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
pynndescent==0.5.13
pyod==2.0.6
pyparsing==3.2.5
pytest==8.4.2
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyWavelets==1.6.0
PyYAML==6.0.3
pyzmq @ file:///opt/miniconda3/conda-bld/pyzmq_1758103938220/work
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
retrying==1.4.2
rich==14.2.0
rich-click==1.9.4
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.7.0
SALib==1.5.1
schemdraw==0.15
scikit-base==0.7.8
scikit-learn==1.6.1
scikit-plot==0.3.7
scipy==1.13.1
seaborn==0.13.2
sentence-transformers==2.2.2
sentencepiece==0.2.1
shap==0.44.1
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
sktime==0.26.0
slicer==0.0.7
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.45
sqlparse==0.5.4
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
starlette==0.49.3
statsforecast==1.5.0
statsmodels==0.14.6
sympy==1.14.0
tbats==1.1.3
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.19.1
tomli==2.3.0
torch==2.8.0
torchvision==0.23.0
tornado @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_63s3osc0_1/croot/tornado_1748956943583/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
transformers==4.40.2
triad==1.0.0
tsdownsample==0.1.4.1
typeguard==4.4.4
typer==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
umap-learn==0.5.7
urllib3==2.6.1
uuid6==2025.0.1
uvicorn==0.38.0
uvloop==0.22.1
visions==0.8.1
waitress==3.0.2
watchdog==6.0.0
watchfiles==1.1.1
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
websockets==15.0.1
Werkzeug==3.1.4
widgetsnbextension==4.0.15
wordcloud==1.9.4
WTForms==3.2.1
wurlitzer==3.1.1
xgboost==2.1.4
xxhash==3.6.0
ydata-profiling==4.17.0
yellowbrick==1.5
zipp==3.23.0
zope.event==6.0
zope.interface==8.0.1
